Added the 'help' command.
Added the help command which can optionally take over the role of the -h
command for individual commands, freeing up that flag for other uses.
When used without arguments, will display the usage string.  When used
with the name of a command, will display the usage string of that
command.

Also added an prearg override for use with the help command to avoid
requiring a prearg when display command usages.